pinctrl: meson-axg: properly handle gpios spying on functions
the meson-axg-pmx driver is not marked as 'strict', meaning it allows
the concurrent use of pinmux functions and gpios. The HW input allows
that for GPIO inputs which are always connected to the pad. The output
driving part of GPIO is not connected when a function other than gpio
is enabled.

While the pinctrl framework will let it happen, this meson-axg-pmx
currently enable the mode regardless of how the pin have already been
claimed, meaning GPIO mode would replace the pre-set function.

To enable gpio input spying of a function, only set the pinmux to gpio mode
if the mux as not already been claimed.

In case a pin is used by a device (A) as function and another device (B) as
GPIO:
* If A probes first, mux is claimed as set. B will get the without altering
  the pinmux
* If B probes first, mux is not claimed but set as gpio. When A probes,
  the GPIO mode is replaced by the function mode as intended.

Maybe this pmx driver should set the mux_owner it GPIO is in output mode.

Fixes: 0fabe43f3f60 ("pinctrl: meson-axg: Introduce a pinctrl pinmux ops for Meson-AXG SoC")
Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>