drm/meson: dw-hdmi: power up phy on device init
The phy is not in a useful state on init.

Other HW, such as eARC, need this phy in working order, even if a mode is
not set on HDMI. This may happen if the display booted on DSI and HDMI mode
has never been setup. In addition, some of the setup done on mode change is
not very has the bits set were already set on device init.

Clean this up and migrate the necessery setup to device init

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
Upstream-Status: Submitted [https://lore.kernel.org/r/20240426160256.3089978-2-jbrunet@baylibre.com]