drm/meson: dw-hdmi: add bandgap setting for g12
When no mode is set (or when the PHY is disabled) the utility pin
appears to be grounded. This is problematic because ARC and eARC use this
line and may do so even if no display mode is set.

This change enable the bandgap setting on g12 chip which fix the problem
with the utility pin. This is done be restoring init values on PHY init
and disable. It will also help with the dw-hdmi rework to get rid of hhi
glabal access and components.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
Upstream-Status: Submitted [https://lore.kernel.org/r/20240426160256.3089978-3-jbrunet@baylibre.com]