drm/meson: dw-hdmi: convert to regmap
The Amlogic mixes direct register access and regmap ones, with several
custom helpers. Using a single API makes rework and maintenance easier.

Convert the Amlogic phy driver to regmap and use it to have more consistent
access to the registers in the driver, with less custom helpers. Add
register bit definitions when missing.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>