ASoC: meson: spdifin: improve stream handling
The amlogic handling of the stream used to be fragile.
The dai allows the usual spdif rates but did not set any constraints based
on the actual input stream because detecting the input is difficult to get
right. The user was expected to check the rate locked through the related
kcontrol, and start the capture accordingly. This could lead to
inconsistencies between the stream rate and capture rate.

Using work done for eARC support, a solution to properly monitor the rate
has been found. With this change, the capture is only allowed to start
if capture rate matches the stream. The capture is terminated if the stream
stops or the rate changes.

Jack support has been added to properly notify userspace when a stream is
present and capture is ready to start.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>