hwmon: pmbus: add tps25990 efuse support
This patchset adds initial support for the Texas Instruments TPS25990
eFuse. The TPS25990 is an integrated, high-current circuit protection and
power management device. TPS25895 may be stacked on the TPS25990 for
higher currents.

This patchset provides basic telemetry support for the device.
On boot, the device is write protected. Limits can be changed in sysfs
if the write protection is removed using the introduced pmbus parameter.

Limits will be restored to the default value device on startup, unless
saved to NVM. Writing the NVM is not supported by the driver at the moment.

As part of this series, PMBus regulator support is improved to better
support write-protected devices.

Changes in v4:
- Rebased on v6.13-rc1, dropping applied patches.
- Replace 0xff with -1 when using pmbus_read/write_byte()
- TPS25990 Support only WP_ALL or 0 as write protection, return -EINVAL otherwise.
- Link to v3: https://lore.kernel.org/r/20241024-tps25990-v3-0-b6a6e9d4b506@baylibre.com

Changes in v3:
- Grouped hwmon write protect patches from:
  https://lore.kernel.org/r/20240920-pmbus-wp-v1-0-d679ef31c483@baylibre.com
- Link to v2: https://lore.kernel.org/r/20240920-tps25990-v2-0-f3e39bce5173@baylibre.com

Changes in v2:
- Drop PGOOD command support
- Use micro-ohms for rimon property and better handle range.
- Adjust read/write callbacks to let PMBus core do the job by default
- Drop history reset specific properties and remap to the generic ones
- Drop debugfs write_protect property and remap to the generic register
- Link to v1: https://lore.kernel.org/r/20240909-tps25990-v1-0-39b37e43e795@baylibre.com

To: Jean Delvare <jdelvare@suse.com>
To: Guenter Roeck <linux@roeck-us.net>
To: Jonathan Corbet <corbet@lwn.net>
To: Jerome Brunet <jbrunet@baylibre.com>
To: Delphine CC Chiu <Delphine_CC_Chiu@Wiwynn.com>
Cc: linux-hwmon@vger.kernel.org
Cc: linux-doc@vger.kernel.org
Cc: linux-kernel@vger.kernel.org
Cc: linux-i2c@vger.kernel.org


--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 4,
    "change-id": "20240909-tps25990-34c0cff2be06",
    "prefixes": [],
    "history": {
      "v1": [
        "20240909-tps25990-v1-0-39b37e43e795@baylibre.com"
      ],
      "v2": [
        "20240920-tps25990-v2-0-f3e39bce5173@baylibre.com"
      ],
      "v3": [
        "20241024-tps25990-v3-0-b6a6e9d4b506@baylibre.com"
      ]
    },
    "prerequisites": [
      "change-id: 20240920-regulator-ignored-data-78e7a855643e:v2",
      "base-commit: 516ddbfef736c843866a0b2db559ce89b40ce378"
    ]
  }
}